requirements.txt
cairocffi==1.3.0
CairoSVG==2.5.2
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
cloudscraper==1.2.60
cssselect2==0.6.0
defusedxml==0.7.1
idna==3.3
Pillow==9.1.0
pycparser==2.21
pyparsing==3.0.8
PyTextNow==1.2.1
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.27.1
requests-toolbelt==0.9.1
six==1.16.0
tinycss2==1.1.1
urllib3==1.26.9
webencodings==0.5.1
